OS/FreeRTOS: Fix compiler warning reported by clang
../../../OS/FreeRTOS/Source/portable/MemMang/heap_4.c:445:46: warning: implicit conversion from 'TickType_t' (aka 'unsigned long long') to 'size_t' (aka 'unsigned int') changes value from 18446744073709551615 to 4294967295 [-Wconstant-conversion]
  445 | size_t xBlocks = 0, xMaxSize = 0, xMinSize = portMAX_DELAY; /* portMAX_DELAY used as a portable way of getting the maximum value. */
      |                                   ~~~~~~~~   ^~~~~~~~~~~~~
../../../OS/FreeRTOS/Source/portable/portmacro.h:68:33: note: expanded from macro 'portMAX_DELAY'
   68 | #define portMAX_DELAY           ( TickType_t )0xFFFFFFFFFFFFFFFFULL
      |                                 ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
1 warning generated.

+++ OS/FreeRTOS/Source/portable/MemMang/heap_4.c
@@ -442,7 +442,7 @@ uint8_t *puc;
 void vPortGetHeapStats( HeapStats_t *pxHeapStats )
 {
 BlockLink_t *pxBlock;
-size_t xBlocks = 0, xMaxSize = 0, xMinSize = portMAX_DELAY; /* portMAX_DELAY used as a portable way of getting the maximum value. */
+size_t xBlocks = 0, xMaxSize = 0, xMinSize = (size_t)(portMAX_DELAY); /* portMAX_DELAY used as a portable way of getting the maximum value. */
 
 	vTaskSuspendAll();
 	{
